Print just the first 25 labels
Some variables have a lot of labels. We were running in a bug where the
ado would quit because of too many values. I now just print the first 25
labels. I might need to add a labels sheet in the future to have a full
print out of the labels.